@@ -15,7 +15,7 @@
 // $URL$
 // $Id$
 // SPDX-License-Identifier: GPL-3.0+
-// 
+//
 //
 // Author(s)     : Fernando Cacciola <Fernando.Cacciola@geometryfactory.com>
 //                 Ronnie Gandhi <ronniegandhi19999@gmail.com>
@@ -31,7 +31,7 @@
 
 #include <QPolygonF>
 #include <QPointF>
-#include <QGraphicsLineItem> 
+#include <QGraphicsLineItem>
 #include <QGraphicsScene>
 #include <QGraphicsSceneMouseEvent>
 
@@ -55,11 +55,11 @@ namespace Qt {
    // typedef typename Bezier_traits::General_polygon_2     Bezier_polygon;
     typedef typename Traits::Rat_kernel::Vector_2         Vector ;
     typedef typename Traits::Rat_kernel::Point_2          Point ;
-    
+
     typedef std::vector<Bezier_curve> Bezier_curve_vector ;
-    
+
     typedef typename Bezier_curve_vector::const_iterator const_bezier_curve_iterator ;
-    
+
     typedef Bezier_boundary_pieces_graphics_item<Bezier_curve_vector> GI ;
 
     GraphicsViewBezierPolygonInput(QObject* aParent, QGraphicsScene* aScene)
@@ -78,80 +78,80 @@ namespace Qt {
       mOngoingPieceGI = new GI(&mOngoingPieceCtr) ;
       mHandle0GI      = new QGraphicsLineItem();
       mHandle1GI      = new QGraphicsLineItem();
-      
+
       mOngoingPieceGI->setPen(mOngoingCurvePen);
       mHandle0GI     ->setPen(mHandlePen);
       mHandle1GI     ->setPen(mHandlePen);
-      
+
       mHandle0GI->setLine(0,0,1,1);
       mHandle1GI->setLine(0,0,1,1);
       mHandle0GI->hide();
       mHandle1GI->hide();
-      
+
       mBezierGI = new GI(&mBezierPolygonPieces) ;
-      
+
       mBezierGI->setPen(mBezierPolygonPen);
-      
+
       mScene->addItem(mOngoingPieceGI);
       mScene->addItem(mHandle0GI);
       mScene->addItem(mHandle1GI);
       mScene->addItem(mBezierGI);
     }
-    
+
     ~GraphicsViewBezierPolygonInput(){}
-    
+
     bool eventFilter(QObject *obj, QEvent *aEvent)
     {
       bool rHandled = false ;
-      
-      if (aEvent->type() == QEvent::GraphicsSceneMousePress) 
+
+      if (aEvent->type() == QEvent::GraphicsSceneMousePress)
       {
         rHandled = mousePressEvent( static_cast<QGraphicsSceneMouseEvent *>(aEvent) ) ;
-      } 
-      else if (aEvent->type() == QEvent::GraphicsSceneMouseRelease) 
+      }
+      else if (aEvent->type() == QEvent::GraphicsSceneMouseRelease)
       {
         rHandled = mouseReleaseEvent( static_cast<QGraphicsSceneMouseEvent *>(aEvent) ) ;
-      } 
-      else if (aEvent->type() == QEvent::GraphicsSceneMouseMove) 
+      }
+      else if (aEvent->type() == QEvent::GraphicsSceneMouseMove)
       {
         rHandled = mouseMoveEvent( static_cast<QGraphicsSceneMouseEvent *>(aEvent) ) ;
-      } 
-      else if (aEvent->type() == QEvent::KeyPress) 
+      }
+      else if (aEvent->type() == QEvent::KeyPress)
       {
         rHandled = keyPressEvent( static_cast<QKeyEvent *>(aEvent) ) ;
       }
-      
+
       if ( !rHandled )
         rHandled = QObject::eventFilter(obj, aEvent);
-              
-      return rHandled ;  
+
+      return rHandled ;
     }
-    
+
   public:
 
     enum State { Start, PieceOrFirstHandleStarted, PieceOngoing, FirstHandleOngoing, HandleOngoing, PieceEnded, CurveEnded } ;
-    
+
     Point cvt ( QPointF const& aP ) const { return Point(aP.x(),aP.y()) ; }
 
     bool mousePressEvent(QGraphicsSceneMouseEvent *aEvent)
     {
-      
+
       bool rHandled = false;
       m_bound_rect = false;
-      
+
       Point lP = cvt(aEvent->scenePos());
 
       if ( aEvent->button() == ::Qt::LeftButton )
       {
         switch (mState)
         {
-          case Start: 
+          case Start:
             mP0      = lP;
             mState   = PieceOrFirstHandleStarted;
             rHandled = true;
             break;
 
-          case PieceOngoing: 
+          case PieceOngoing:
             mP1      = lP;
             mState   = HandleOngoing;
             rHandled = true;
@@ -173,7 +173,7 @@ namespace Qt {
            default: break; //! \todo handle default case
          }
     }
-      
+
       return rHandled ;
     }
 
@@ -181,15 +181,15 @@ namespace Qt {
     {
       bool rHandled = false ;
       Point lP = cvt(aEvent->scenePos());
-      
+
       switch (mState)
       {
         case PieceOrFirstHandleStarted:
           mState   = FirstHandleOngoing;
           rHandled = true;
           break;
-          
-        case PieceOngoing: 
+
+        case PieceOngoing:
           mP1 = lP;
           UpdateOngoingPiece();
           rHandled = true ;
@@ -199,7 +199,7 @@ namespace Qt {
           UpdateVeryFirstHandle(lP);
           rHandled = true ;
           break;
-          
+
         case HandleOngoing:
           if(m_last)
           {
@@ -210,22 +210,22 @@ namespace Qt {
           UpdateOngoingPiece();
           rHandled = true ;
           break;
-          
+
         case PieceEnded:
           mState   = PieceOngoing;
           rHandled = true;
           break;
       }
-      
+
       return rHandled ;
     }
-    
+
     bool mouseReleaseEvent(QGraphicsSceneMouseEvent *aEvent)
     {
       bool rHandled = false ;
-      
+
       Point lP = cvt(aEvent->scenePos());
-      
+
       if ( aEvent->button() == ::Qt::LeftButton )
       {
         switch (mState)
@@ -234,16 +234,16 @@ namespace Qt {
             mState   = PieceOngoing;
             rHandled = true;
             break;
-            
+
           case FirstHandleOngoing:
             UpdateVeryFirstHandle(lP);
             mPrevH0  = mH1 ;
-            mH1      = boost::optional<Point>();
+            mH1      = std::optional<Point>();
             mState   = PieceOngoing;
             rHandled = true;
             break;
-          
-          case HandleOngoing: 
+
+          case HandleOngoing:
             UpdateHandles(lP);
             CommitOngoingPiece(lP);
             mState   = PieceEnded;
@@ -257,10 +257,10 @@ namespace Qt {
       {
         switch (mState)
         {
-          case HandleOngoing: 
+          case HandleOngoing:
             m_bound_rect = false;
             if(m_last_bezier)
-            { 
+            {
               HideHandles();
               CommitOngoingPiece(lP);
             }
@@ -271,17 +271,17 @@ namespace Qt {
             break;
 
             default: break; //!todo handle default case
-        }    
+        }
       }
       return rHandled ;
     }
-    
+
     bool keyPressEvent(QKeyEvent *aEvent)
     {
       bool rHandled = false ;
-      
+
       if( aEvent->key() == ::Qt::Key_Delete || aEvent->key() == ::Qt::Key_Backspace )
-      {     
+      {
         RemoveLastPiece();
         mState   = mBezierPolygonPieces.size() > 0 ? PieceEnded : Start ;
         rHandled = true;
@@ -302,10 +302,10 @@ namespace Qt {
 
     void ReStart()
     {
-      mPrevH0 = mH0 = mH1 = boost::optional<Point>();
-      mState = Start ;     
+      mPrevH0 = mH0 = mH1 = std::optional<Point>();
+      mState = Start ;
     }
-    
+
     void Reset()
     {
       mBezierPolygonPieces.clear();
@@ -314,7 +314,7 @@ namespace Qt {
       mOngoingPieceGI->modelChanged();
       ReStart();
     }
-    
+
     void RemoveLastPiece()
     {
       mBezierPolygonPieces.pop_back();
@@ -326,14 +326,14 @@ namespace Qt {
         mP0 = mBezierPolygonPieces.back().control_point(mBezierPolygonPieces.back().number_of_control_points()-1);
         UpdateOngoingPiece();
       }
-      mPrevH0 = mH0 = mH1 = boost::optional<Point>();
-    }      
+      mPrevH0 = mH0 = mH1 = std::optional<Point>();
+    }
 
     void HideHandles()
     {
       mHandle0GI->hide();
       mHandle1GI->hide();
-    }  
+    }
 
     Bezier_curve CreatePiece()
     {
@@ -342,7 +342,7 @@ namespace Qt {
         Point lControlPoints[4] = { mP0
                                   , *mPrevH0
                                   , *mH1
-                                  , mP1 
+                                  , mP1
                                   } ;
         return Bezier_curve( lControlPoints, lControlPoints + 4 ) ;
       }
@@ -350,7 +350,7 @@ namespace Qt {
       {
         Point lControlPoints[3] = { mP0
                                   , *mPrevH0
-                                  , mP1 
+                                  , mP1
                                   } ;
         return Bezier_curve( lControlPoints, lControlPoints + 3 ) ;
       }
@@ -358,30 +358,30 @@ namespace Qt {
       {
         Point lControlPoints[3] = { mP0
                                   , *mH1
-                                  , mP1 
+                                  , mP1
                                   } ;
         return Bezier_curve( lControlPoints, lControlPoints + 3 ) ;
       }
       else
       {
         Point lControlPoints[2] = { mP0
-                                  , mP1 
+                                  , mP1
                                   } ;
         return Bezier_curve( lControlPoints, lControlPoints + 2 ) ;
       }
     }
-    
+
     void UpdateOngoingPiece()
     {
       if ( mOngoingPieceCtr.size() > 0 )
            mOngoingPieceCtr.clear();
            mOngoingPieceCtr.push_back(CreatePiece());
            mOngoingPieceGI->modelChanged();
-    }      
-    
+    }
+
     void CommitOngoingPiece( Point const& aP )
     {
-      if ( ongoing_piece() ) 
+      if ( ongoing_piece() )
       {
         mBezierPolygonPieces.push_back( *ongoing_piece() ) ;
         mBezierGI->modelChanged();
@@ -390,9 +390,9 @@ namespace Qt {
         mP0 = mP1 ;
         mP1 = aP ;
         mPrevH0 = mH0 ;
-        mH0 = mH1 = boost::optional<Point>();
+        mH0 = mH1 = std::optional<Point>();
       }
-    }      
+    }
 
     void UpdateVeryFirstHandle(Point const& aP)
     {
@@ -403,13 +403,13 @@ namespace Qt {
                              to_double(mH1->x()), to_double(mH1->y()));
         mHandle1GI->show();
 
-        mH0 = boost::optional<Point>();
+        mH0 = std::optional<Point>();
         mHandle0GI->hide();
       }
       else
       {
         HideHandles();
-        mH0 = mH1 = boost::optional<Point>();
+        mH0 = mH1 = std::optional<Point>();
       }
     }
 
@@ -419,7 +419,7 @@ namespace Qt {
       {
         mH0 = aP ;
         mH1 = mP1 - (aP - mP1);
-        
+
         mHandle0GI->setLine( to_double(mP1.x()), to_double(mP1.y()), to_double(mH0->x()), to_double(mH0->y()));
         mHandle1GI->setLine( to_double(mP1.x()), to_double(mP1.y()), to_double(mH1->x()), to_double(mH1->y()));
         mHandle0GI->show();
@@ -428,71 +428,70 @@ namespace Qt {
       else
       {
         HideHandles();
-        mH0 = mH1 = boost::optional<Point>();
+        mH0 = mH1 = std::optional<Point>();
       }
     }
-        
+
     void CloseCurrBundary()
     {
       if ( mBezierPolygonPieces.size() > 0 && ongoing_piece()!= NULL  && !m_last_bezier)
       {
         std::vector<Point> lControlPoints(ongoing_piece()->control_points_begin(),ongoing_piece()->control_points_end());
-        
+
         lControlPoints.back() = mBezierPolygonPieces.front().control_point(0);
-              
+
         mBezierPolygonPieces.push_back( Bezier_curve( lControlPoints.begin(), lControlPoints.end() ) ) ;
-        
+
         mBezierGI->modelChanged() ;
       }
     }
-        
+
     void CommitCurrBezierPolygon()
     {
       GenerateBezierPolygon();
 
       mOngoingPieceCtr.clear();
       mOngoingPieceGI->modelChanged();
-      
+
       mBezierPolygonPieces.clear();
       mBezierGI->modelChanged() ;
-      
-      mPrevH0 = mH0 = mH1 = boost::optional<Point>();
-      
+
+      mPrevH0 = mH0 = mH1 = std::optional<Point>();
+
       HideHandles();
     }
-    
-    void GenerateBezierPolygon() 
-    {
-      Traits traits ;
-      typename Traits::Make_x_monotone_2 make_x_monotone = traits.make_x_monotone_2_object();
-      typedef boost::variant<Bezier_X_monotone_curve ,Point> Make_x_monotone_result;
-      
+
+    void GenerateBezierPolygon() {
+      Traits traits;
+      auto make_x_monotone = traits.make_x_monotone_2_object();
+      using Make_x_monotone_result =
+        std::variant<typename Traits::Point_2, Bezier_X_monotone_curve>;
+
       std::vector<Bezier_X_monotone_curve> xcvs;
 
-      for ( const_bezier_curve_iterator it = mBezierPolygonPieces.begin() ; it != mBezierPolygonPieces.end() ; ++ it )
-      {
-        std::vector<Make_x_monotone_result>       x_objs;
-        make_x_monotone ( *it, std::back_inserter (x_objs));
-        for(auto i=0;i<x_objs.size();++i)
-        {
-            auto* xcv = boost::get<Bezier_X_monotone_curve>(&x_objs[i]);
+      for (auto it = mBezierPolygonPieces.begin();
+           it != mBezierPolygonPieces.end(); ++ it) {
+        std::vector<Make_x_monotone_result> x_objs;
+        make_x_monotone(*it, std::back_inserter(x_objs));
+        for (auto i = 0; i < x_objs.size(); ++i) {
+            auto* xcv = std::get_if<Bezier_X_monotone_curve>(&x_objs[i]);
             CGAL_assertion(xcv != nullptr);
             xcvs.push_back(*xcv);
         }
       }
-      
+
       if ( xcvs.size() > 0 )
       {
         m_last = false;
         m_last_bezier = false;
         Bezier_polygon bp(xcvs.begin(), xcvs.end());
-        emit(generate(boost::variant<std::pair<Bezier_polygon,Bezier_curve_vector> >(std::make_pair(bp,mBezierPolygonPieces))));
-      }  
+        emit(generate(std::variant<std::pair<Bezier_polygon, Bezier_curve_vector> >(std::make_pair(bp,mBezierPolygonPieces))));
+      }
     }
 
     void get_BoundingRect()
     {
-     
+
       m_bound_rect = true;
 
       mP0 = Point(-15500000,-10000000);
@@ -541,34 +540,34 @@ namespace Qt {
     {
       return m_bound_rect;
     }
-    
+
   public:
     QGraphicsScene*    mScene ;
-    GI*                mBezierGI ; 
-    GI*                mOngoingPieceGI ; 
-    QGraphicsLineItem* mHandle0GI ;          
-    QGraphicsLineItem* mHandle1GI ;          
+    GI*                mBezierGI ;
+    GI*                mOngoingPieceGI ;
+    QGraphicsLineItem* mHandle0GI ;
+    QGraphicsLineItem* mHandle1GI ;
 
     QPen mBezierPolygonPen ;
     QPen mOngoingCurvePen ;
-    QPen mHandlePen ;    
+    QPen mHandlePen ;
 
     bool m_bound_rect;
     bool m_last_bezier;
     bool m_last;
-      
+
     Bezier_curve_vector mBezierPolygonPieces ;
-    Bezier_curve_vector mOngoingPieceCtr ;  
+    Bezier_curve_vector mOngoingPieceCtr ;
 
     int mState;
-    
+
     Point mP0;
     Point mP1;
-    
-    boost::optional<Point> mPrevH0;
-    boost::optional<Point> mH0;
-    boost::optional<Point> mH1;
-  
+
+    std::optional<Point> mPrevH0;
+    std::optional<Point> mH0;
+    std::optional<Point> mH1;
+
   }; // end class GraphicsViewBezierPolygonInput
 
 } // namespace Qt
